requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.2.1
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
iniconfig==2.1.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
uvicorn==0.35.0
Werkzeug==3.1.3
whitenoise==6.9.0
